+++ elegy/nn/__init__.py
@@ -1,0 +1,5 @@
+from .batch_normalization import BatchNormalization
+from .dropout import Dropout
+from .flatten import Flatten
+from .linear import Linear
+from .sequential import Sequential