+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## [0.1.4] - 2020-07-24
+* Adds `elegy.metrics.BinaryCrossentropy` @sebasarango1180
+* Adds `elegy.nn.Dropout` and `elegy.nn.BatchNormalization` @cgarciae
+* Improves documentation
+* Fixes bug that cause error when using `is_training` via dependency injection on `Model.predict`.
+
 ## [0.1.3] - 2020-07-22
 * Initial release